requirements.txt
asgiref==3.4.1
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
Django==4.0
django-3-jet==1.0.8
django-admin-interface==0.18.3
django-colorfield==0.6.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jazzmin==2.4.8
django-redis==5.2.0
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
packaging==21.3
passlib==1.7.4
Pillow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytz==2021.3
redis==4.1.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
whitenoise==5.3.0
wrapt==1.13.3
zope.interface==5.4.0
